# Optimize real matrix * complex matrix by using real matmul