--- a/stdlib/LinearAlgebra/src/matmul.jl
+++ b/stdlib/LinearAlgebra/src/matmul.jl
@@ -174,12 +174,8 @@ function (*)(A::AdjOrTransStridedMat{<:BlasComplex}, B::StridedMaybeAdjOrTransMa
          wrapperop(B)(convert(AbstractArray{real(TS)}, _parent(B))))
 end
 # the following case doesn't seem to benefit from the translation A*B = (B' * A')'
-function (*)(A::StridedMatrix{<:BlasReal}, B::StridedMatrix{<:BlasComplex})
-    TS = promote_type(eltype(A), eltype(B))
-    mul!(similar(B, TS, (size(A, 1), size(B, 2))),
-         convert(AbstractArray{TS}, A),
-         convert(AbstractArray{TS}, B))
-end
+(*)(A::StridedMatrix{<:BlasReal}, B::StridedMatrix{<:BlasComplex}) =
+    (A * real(B)) .+ im .* (A * imag(B))
 (*)(A::AdjOrTransStridedMat{<:BlasReal}, B::StridedMatrix{<:BlasComplex}) = copy(transpose(transpose(B) * parent(A)))
 (*)(A::StridedMaybeAdjOrTransMat{<:BlasReal}, B::AdjOrTransStridedMat{<:BlasComplex}) = copy(wrapperop(B)(parent(B) * transpose(A)))