--- a/stdlib/LinearAlgebra/src/matmul.jl
+++ b/stdlib/LinearAlgebra/src/matmul.jl
@@ -175,7 +175,7 @@ function (*)(A::AdjOrTransStridedMat{<:BlasComplex}, B::StridedMaybeAdjOrTransMa
 end
 # the following case doesn't seem to benefit from the translation A*B = (B' * A')'
 (*)(A::StridedMatrix{<:BlasReal}, B::StridedMatrix{<:BlasComplex}) =
-    (A * real(B)) .+ im .* (A * imag(B))
+    Complex.(A * real(B), A * imag(B))
 (*)(A::AdjOrTransStridedMat{<:BlasReal}, B::StridedMatrix{<:BlasComplex}) = copy(transpose(transpose(B) * parent(A)))
 (*)(A::StridedMaybeAdjOrTransMat{<:BlasReal}, B::AdjOrTransStridedMat{<:BlasComplex}) = copy(wrapperop(B)(parent(B) * transpose(A)))